Remove support for automatic token renewal
Auto-renewal seems to be the cause for the flooding that we've been
experiencing in our spotify refresh endpoints. On this commit, we're
disabling it since it's not really needed for our use case. The requests
that we send to the spotify sdk will always confirm if the tokens are
still valid before sending requests (refreshing if not).